pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = []
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
dj-database-url = "^2.3.0"
django-celery-results = "^2.5.1"
django-cors-headers = "^4.6.0"
django-environ = "^0.11.2"
drf-spectacular = "^0.27.2"
drf-standardized-errors = {extras = ["openapi"], version = "^0.14.1"}
httpx = "^0.27.2"
web3 = "^7.6.0"
siwe = "^4.4.0"
pydantic = "^2.10.1"
gevent = "^24.11.1"
gunicorn = "^23.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
pytz = "^2024.2"
whitenoise = "^6.8.2"
django-structlog = {extras = ["celery"], version = "^8.1.0"}
rich = "^13.9.4"
django-rest-knox = "^5.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
django-stubs = "^5.1.1"
djangorestframework-stubs = "^3.15.1"
pre-commit = "^4.0.1"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--showlocals --no-migrations"
testpaths = ["tests"]
filterwarnings = [
    "ignore:datetime.datetime.utcnow*:DeprecationWarning",
]
DJANGO_SETTINGS_MODULE = "api.settings.testing"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
extend-select = ["I"] # enable import sorting
ignore = ["E266", "E501", "F403", "W291"]

[tool.django-stubs]
django_settings_module = "api.settings.local"